+++ TodoTxtDaemon.UnitTests/WorkerTests.cs
@@ -6,8 +6,6 @@ namespace TodoTxtDaemon.UnitTests
     {
         private readonly Mock<ILogger<Worker>> _LoggerMock;
 
-        private readonly Mock<IHostEnvironment> _HostEnvironmentMock;
-
         private readonly Mock<IHostApplicationLifetime> _LifetimeMock;
 
         private readonly Mock<IWatcher> _WatcherMock;
@@ -19,22 +17,17 @@ namespace TodoTxtDaemon.UnitTests
         public WorkerTests()
         {
             _LoggerMock = new Mock<ILogger<Worker>>(MockBehavior.Strict);
-            _HostEnvironmentMock = new Mock<IHostEnvironment>(MockBehavior.Strict);
-            _LifetimeMock = new Mock<IHostApplicationLifetime>(MockBehavior.Strict);
-            _WatcherMock = new Mock<IWatcher>(MockBehavior.Strict);
-            _MoverMock = new Mock<IMover>(MockBehavior.Strict);
-            _Worker = new Worker(_LoggerMock.Object, _HostEnvironmentMock.Object,
-                _LifetimeMock.Object, _WatcherMock.Object, _MoverMock.Object);
             _LoggerMock.Setup(LogLevel.Information);
             _LoggerMock.Setup(LogLevel.Error);
             _LoggerMock.Setup(LogLevel.Critical);
-            _HostEnvironmentMock.Setup(h => h.ContentRootPath).Returns("/test/path");
-            _LifetimeMock.Setup(l => l.ApplicationStopped)
-                .Returns(CancellationToken.None);
+            _LifetimeMock = new Mock<IHostApplicationLifetime>(MockBehavior.Strict);
             _LifetimeMock.Setup(l => l.StopApplication());
+            _WatcherMock = new Mock<IWatcher>(MockBehavior.Strict);
             _WatcherMock.Setup(w => w.IsTimeToRun()).Returns(true);
             _WatcherMock.Setup(w => w.MarkRun());
+            _MoverMock = new Mock<IMover>(MockBehavior.Strict);
             _MoverMock.Setup(m => m.Run());
+            _Worker = new Worker(_LoggerMock.Object, _LifetimeMock.Object, _WatcherMock.Object, _MoverMock.Object);
         }
 
         [Fact]
@@ -70,33 +63,6 @@ namespace TodoTxtDaemon.UnitTests
             _LoggerMock.Verify(LogLevel.Error, "test message");
             _WatcherMock.Verify(w => w.MarkRun(), Times.Once);
             _MoverMock.Verify(m => m.Run(), Times.Once);
-            VerifyNoOtherCalls();
-        }
-
-        [Fact]
-        public async Task ExecuteAsync_Stops_WhenIsTimeToRunThrowsWatcherException()
-        {
-            _WatcherMock.Setup(m => m.IsTimeToRun()).Throws(new WatcherException("test message"));
-
-            await ExecuteAsync();
-
-            VerifyCommonInvocations();
-            _LoggerMock.Verify(LogLevel.Critical, "test message");
-            _LifetimeMock.Verify(l => l.StopApplication(), Times.Once);
-            VerifyNoOtherCalls();
-        }
-
-        [Fact]
-        public async Task ExecuteAsync_Stops_WhenMarkRunThrowsWatcherException()
-        {
-            _WatcherMock.Setup(m => m.MarkRun()).Throws(new WatcherException("test message"));
-
-            await ExecuteAsync();
-
-            VerifyCommonInvocations();
-            _LoggerMock.Verify(LogLevel.Critical, "test message");
-            _LifetimeMock.Verify(l => l.StopApplication(), Times.Once);
-            _WatcherMock.Verify(w => w.MarkRun(), Times.Once);
             VerifyNoOtherCalls();
         }
 
@@ -161,16 +127,13 @@ namespace TodoTxtDaemon.UnitTests
         {
             _LoggerMock.Verify(LogLevel.Information, "Monitoring...", Times.Exactly(monitoringLogCalls));
             var version = typeof(Program).Assembly.GetName().Version?.ToString(3);
-            _LoggerMock.Verify(LogLevel.Information, $"TodoTxtDaemon {version} started. Current working directory: /test/path");
-            _HostEnvironmentMock.Verify(h => h.ContentRootPath, Times.Once);
-            _LifetimeMock.Verify(l => l.ApplicationStopped, Times.Once);
+            _LoggerMock.Verify(LogLevel.Information, $"TodoTxtDaemon {version} started. Process Id: ");
             _WatcherMock.Verify(w => w.IsTimeToRun(), Times.Once);
         }
 
         private void VerifyNoOtherCalls()
         {
             _LoggerMock.VerifyNoOtherCalls();
-            _HostEnvironmentMock.VerifyNoOtherCalls();
             _LifetimeMock.VerifyNoOtherCalls();
             _WatcherMock.VerifyNoOtherCalls();
             _MoverMock.VerifyNoOtherCalls();